+++ CHANGES.md
@@ -23,6 +23,7 @@
   to print
 - `generated_source()` steps are now automatically created as necessary for
   files passed to `object_file()`, if possible
+- Add `path_exists()` builtin
 - Paths with a trailing `/` are now treated as directory paths, and when passed
   to `auto_file()` will create a directory object
 - Add support for compiling Windows `.rc` files